+++ src/main/scala/leo/modules/calculus/NormalizationRules.scala
@@ -985,8 +985,8 @@ object Simp extends CalculusRule {
     if (t.sharing) Term.insert(result)
     else result
   }
-  @tailrec private[this] final def gcd(a: Int, b: Int): Int = if (b == 0) a.abs else gcd(b, a % b)
-  private[this] final def normalizeReal(wholePart: Int, decimalPlaces: Int, exponent: Int): Term = {
+  @tailrec private[this] final def gcd(a: BigInt, b: BigInt): BigInt = if (b == 0) a.abs else gcd(b, a % b)
+  private[this] final def normalizeReal(wholePart: BigInt, decimalPlaces: BigInt, exponent: BigInt): Term = {
 //    val decimalPlacesWithoutTrailingZeroes = if (decimalPlaces != 0) decimalPlaces.toString.reverse.dropWhile(_ == '0').reverse.toInt else 0
 //    val decimalPlacesWithoutTrailingZeroesLength = decimalPlacesWithoutTrailingZeroes.toString.length
 //    val wholePartAsString = wholePart.toString
@@ -1154,8 +1154,8 @@ object Simp extends CalculusRule {
           }
         case Rational(n, d) =>
           assert(args.isEmpty, "Applied arguments to a rational [termSimp0]")
-          val sign: Int = d.sign
-          val greatestCommonDivisor: Int = gcd(n ,d).abs * sign
+          val sign: Int = d.sign.toInt
+          val greatestCommonDivisor: BigInt = gcd(n ,d).abs * sign
           mkRational(n / greatestCommonDivisor, d / greatestCommonDivisor)
         case Real(w,d,e) =>
           assert(args.isEmpty, "Applied arguments to a real [termSimp0]")
