+++ src/main/scala/leo/modules/procedures/Simplification.scala
@@ -79,8 +79,8 @@ object Simplification extends Function1[Term, Term] {
       case ty :::> body => mkTermAbs(ty, apply0(body, extensional))
       case TypeLambda(body) => mkTypeAbs(apply0(body, extensional))
       case Rational(n, d) =>
-        val sign: Int = d.sign
-        val greatestCommonDivisor: Int = gcd(n ,d).abs * sign
+        val sign: Int = d.sign.toInt
+        val greatestCommonDivisor: BigInt = gcd(n ,d).abs * sign
         mkRational(n / greatestCommonDivisor, d / greatestCommonDivisor)
       case Real(w,d,e) => normalizeReal(w,d,e)
       case f ∙ args if f.isConstant && args.length <= 3 =>
@@ -254,8 +254,8 @@ object Simplification extends Function1[Term, Term] {
     }
   }
 
-  @tailrec private[this] final def gcd(a: Int, b: Int): Int = if (b == 0) a.abs else gcd(b, a % b)
-  private[this] final def normalizeReal(wholePart: Int, decimalPlaces: Int, exponent: Int): Term = { // TODO
+  @tailrec private[this] final def gcd(a: BigInt, b: BigInt): BigInt = if (b == 0) a.abs else gcd(b, a % b)
+  private[this] final def normalizeReal(wholePart: BigInt, decimalPlaces: BigInt, exponent: BigInt): Term = { // TODO
     //    val decimalPlacesWithoutTrailingZeroes = if (decimalPlaces != 0) decimalPlaces.toString.reverse.dropWhile(_ == '0').reverse.toInt else 0
     //    val decimalPlacesWithoutTrailingZeroesLength = decimalPlacesWithoutTrailingZeroes.toString.length
     //    val wholePartAsString = wholePart.toString
